Windows 8.1 Audio background task
Adding an audio background task allows the audio to continue working
between 2 peers after the app is sent to the background.
Surprisingly, this is also improving the behavior of video capture
(local video no longer frozen when the application is back in the foreground).

A few issues have been noticed:
- the video may freeze after the app is brought to background and back to
foreground repeatedly
- when the rendering image is shrinking, the local video can be seen
(flickering between green and the live video stream) in the background