Task 9736: PeerCC: need to tap several times to reconnect after hung up
It turned out that one of the webrtc functions taks long time to be executed:
SetInterface(NetworkInterface *iface) in mediachannel.h.

[Suggested solution]
Let the Connect to peer button to be disabled until the connection closure process
is fully completed

[Test Cases]
1. Make sure that after hug up the Connect to Peer button remains disabled until
the peer connection is closure process is fully completed.
2. Make sure that after the button becomes enabled.

[Known issue]
1. After hung up very often the app re-initiates the connection with the peer.
This behaviour existed in the application before this solution is suggested.
Thus it is preferrable to be fixed separately