[Task 9735]: UI Controls adjustments for Win10
[Description and Solution]
The UI of the PeerCC app on Win10 phone is not good. The
Settings page is almost not visible thus the end users cannot
use it.

The solution is not perfect, but at least the user can scroll the
UI to see the controls and use them.
It was decided to go with the easiest and quickest soluion which
is to make the application main page scrollable so the settings
container can be scrolled and used as needed.

[Changes]
MainPage.xaml:
- main content is included in a ScrollViewer (The VSO
shows lot of changes because of changed intents).

[Test Cases]
1. Build and Run the app on Windows 10 desktop and Windows Phone 10
2. Verify that all UI controls are visible
3. Hit the Settings button, verify that the UI is scrollable on the phone
and you can use the settings layout.
4. Close the Setting page on phone -> Verify that page is not scrollable
(The scroll will become available on any device if the entire application
layout is bigger than the device screen, or, if the app windows is resized
on desktop)
5. Start a session with a peer
6. Hung up the call, verify that the page is scrollable when settings is
open.