Task-10367:Fix WP UI layout
[Description]
Connect button remains under the border and is not accessible
while the phone is in landscape view. It becomes reachable only
in vertical mode or when the settings view is open.

[Note from Dev]
The Connect button will be unreachable in the Portrait mode as
well if the list of connected peers is long enough not to fit
the device screen.

[Changes and Solution]
Added a helper function for managing the ScrollBarVisibility type.
The app behavior will be as follows and it will not depand on
device orientation:
Settings is not open, not connected to peer - Auto scrollable UI
Settings is not open, connected to peer - Disabled scroll
Settings is open, not connected to peer - Auto scrollable UI
Settings is open, connected to peer - Auto scrollable UI

[Test Cases]
1. Test the above mentioned cases on Windows and Windows Phone.
2. Regression tests: Verify the changes do not affect the app
behavior on Win8.1 and WP8.1