Bugs 10109, 9697 and 10263
[Description]
1. Crash: on hang up immediately after connecting to a peer
2. Crash: Reconnect to peer immediately after hang up
3. PeerCC trying to re-establish the call after quick hung up

[Solution]
For 1 and 2: Disable the hung up button until the app is ready to
disconnect from the peer (this is just a workaround, as discussed
offline we are going to refactor the Conductor/Sognalling in the
future)
For 2 and 3: When creating a peer connection on OnMessageFromPeer
we need to make sure that the message is for establishing a call. Otherwise
the message could be from old connections which are not fully closed yet
on the current client.

[Test Cases]
1. Run the PeerCC on Win 8.1 desktop/Win Phone 8.1/Windows 10/ Windows
Phone 10
2. Establish a call with a peer
3. Verify the hung up button is disabled until the app is ready to disconnect
from peer, after it becomes enabled
4. Verify you can successfully hung up the call
5. Establish a call and as soon as the hung up button is enabled very
quickly hit to end the call -> Verify the app does not crash and the app
does not try to reconnect back again
6. As soon as the call view is gone very quickly hit the connect button ->
Verify the app is not crashing on connect (it may take time to connect
to the same peer as the remote peer could not be ready yet, you should
wait for some seconds)
Note: The crashes could be reproducible for Disconnect button (initiate a
call and quickly hit the Disconnect button, but it is out of scope of these
bugs)