Added hack for triggering RTMediaStreamSource teardown.
This is a hack/workaround for destroying the internal stream source (RTMediaStreamSource) instance inside webrtc winrt api when loopback is disabled.
For some reason, the RTMediaStreamSource instance is not destroyed when only SelfVideo.Source is set to null.
For unknown reasons, when executing the above sequence (set to null, stop, set to null), the internal stream source is destroyed.
Apparently, with webrtc package version < 1.1.175, the internal stream source was destroyed corectly, only by setting SelfVideo.Source to null.

Validation:
- Start PCC, establish call
- Disabled video loopback
- Make sure RTMediaStreamSource instance associated with video loopback is destroyed (watch in logs for: 'RTMediaStreamSource::~RTMediaStreamSource', or add breakpoint in destructor)